cli: split sessions cli tests out of production module
